fix: fixed types
* doesn't quite fix #13, but typedoc is showing no export errors now
* fixing typing when this package is complied with another dependency which would be needed
* moved unit test functions back into its own file
