+++ src/builder/interface/node.ts
@@ -1,4 +1,7 @@
-/** @internal */
+/**
+ * Node
+ * @since 1.0.0
+ */
 export interface Node {
   name: string
   length: number
