+++ src/builder/modules/emptyNode.ts
@@ -1,6 +1,11 @@
 import { Node } from '../interface/node.js'
 
-/** @internal */
+/**
+ * Empty Node
+ * @description Used for anything not value in a segment index, this will be the defaults of the methods.
+ * Some have to be implemented on those segments to work.
+ * @since 1.0.0
+ */
 export class EmptyNode implements Node {
   get name (): string {
     throw new Error('Method not implemented')
