+++ src/builder/modules/nodeBase.ts
@@ -1,4 +1,4 @@
-import { isNumber, isString, pad } from '../../utils/utils.js'
+import { isHL7Number, isHL7String, padHL7Date } from '../../utils/utils.js'
 import { Batch } from '../batch.js'
 import { EmptyNode } from './emptyNode.js'
 import { HL7FatalError } from '../../utils/exception.js'
@@ -6,7 +6,11 @@ import { Delimiters } from '../../utils/enum.js'
 import { Node } from '../interface/node.js'
 import { Message } from '../message.js'
 
-/** @internal */
+/**
+ * Node Base
+ * @since 1.0.0
+ * @implements Node
+ */
 export class NodeBase implements Node {
   protected parent: NodeBase | null
 
@@ -64,7 +68,7 @@ export class NodeBase implements Node {
       }
 
       return this
-    } else if (isNumber(path)) {
+    } else if (isHL7Number(path)) {
       if (Array.isArray(value)) {
         const child = this.ensure(path)
         for (let i = 0, l = value.length; i < l; i++) {
@@ -156,7 +160,7 @@ export class NodeBase implements Node {
     }
     if (typeof path === 'number') {
       return this.setChild(this.createChild('', path), path)
-    } else if (isString(path)) {
+    } else if (isHL7String(path)) {
       return this.write(this.preparePath(path), '')
     }
     throw new HL7FatalError(500, 'There seems to be a problem.')
@@ -335,12 +339,12 @@ export class NodeBase implements Node {
   private _formatDateTime (date: Date): string {
     // check if there is a time component
     if (date.getHours() !== 0 || date.getMinutes() !== 0 || date.getSeconds() !== 0 || date.getMilliseconds() !== 0) {
-      return this._formatDate(date) + pad(date.getHours(), 2) + pad(date.getMinutes(), 2) + pad(date.getSeconds(), 2)
+      return this._formatDate(date) + padHL7Date(date.getHours(), 2) + padHL7Date(date.getMinutes(), 2) + padHL7Date(date.getSeconds(), 2)
     }
     return this._formatDate(date)
   }
 
   private _formatDate (date: Date): string {
-    return date.getFullYear().toString() + pad(date.getMonth() + 1, 2) + pad(date.getDate(), 2)
+    return date.getFullYear().toString() + padHL7Date(date.getMonth() + 1, 2) + padHL7Date(date.getDate(), 2)
   }
 }
