+++ src/index.ts
@@ -1,6 +1,5 @@
 import { HL7_2_7, HL7_2_7_BHS, HL7_2_7_FHS } from './specification/2.7.js'
 import { HL7_SPEC, HL7_SPEC_BASE } from './specification/specification.js'
-
 import { NodeBase } from './builder/modules/nodeBase.js'
 import { EmptyNode } from './builder/modules/emptyNode.js'
 import { Node } from './builder/interface/node.js'
@@ -18,18 +17,18 @@ import { ParserPlan } from './utils/parserPlan.js'
 import { HL7Outbound, OutboundHandler } from './client/hl7Outbound.js'
 import { Delimiters, ReadyState } from './utils/enum.js'
 
-export { sleep, expectEvent, expBackoff, createDeferred, Deferred, assertNumber, isNumber, isString, validIPv4, validIPv6, createHL7Date, isBatch, isFile, pad, escapeForRegExp, decodeHexString, randomString } from './utils/utils.js'
+export { expBackoff, assertNumber, isHL7Number, isHL7String, validIPv4, validIPv6, createHL7Date, isBatch, isFile, padHL7Date, escapeForRegExp, decodeHexString, randomString } from './utils/utils.js'
 
 /** HL7 Specs **/
-export type { MSH } from './specification/specification.js'
-export type { HL7_2_7_MSH } from './specification/2.7.js'
+export { MSH, BHS, FHS } from './specification/specification.js'
+export { HL7_2_7_MSH } from './specification/2.7.js'
 
 /** HL7 Class **/
-export { HL7_2_7_BHS, HL7_2_7_FHS, HL7_SPEC, HL7_SPEC_BASE, HL7_2_7 }
+export { HL7_2_7_BHS, HL7_2_7_FHS, HL7_2_7, HL7_SPEC, HL7_SPEC_BASE }
 
 export type { ClientOptions, ClientListenerOptions } from './utils/normalizedClient.js'
 export type { ClientBuilderFileOptions, ClientBuilderBatchOptions, ClientBuilderMessageOptions, ClientBuilderOptions } from './utils/normalizedBuilder.js'
 export type { HL7Error, HL7FatalError, HL7ParserError } from './utils/exception.js'
 
 export default Client
-export { Client, HL7Outbound, OutboundHandler, ParserPlan, FileBatch, Batch, Message, Segment, SegmentList, Component, SubComponent, Field, FieldRepetition, NodeBase, EmptyNode, Node, Delimiters, ReadyState }
+export { Client, HL7Outbound, OutboundHandler, FileBatch, Batch, Message, Segment, SegmentList, Component, SubComponent, Field, FieldRepetition, ParserPlan, Node, NodeBase, EmptyNode, Delimiters, ReadyState }
